+++ R/ipeds_completions.R
@@ -46,7 +46,7 @@ get_grad_rates <- function(UNITIDs = NULL){
   # expects ef<Year>d.csv
 
   # find all the tables
-  tnames <- my_dbListTables(search_string = "^GR20\\d\\d$")
+  tnames <- my_dbListTables(search_string = "^gr20\\d\\d$")
 
   out <- data.frame()
 
@@ -132,7 +132,7 @@ get_grad_demo_rates <- function(UNITIDs = NULL){
   }
 
   # find all the tables
-  tnames <- my_dbListTables(search_string = "^GR20\\d\\d$")
+  tnames <- my_dbListTables(search_string = "^gr20\\d\\d$")
 
   out <- data.frame()
 
@@ -202,7 +202,7 @@ get_grad_pell_rates <- function(UNITIDs = NULL){
   }
 
   # find all the tables
-  tnames <- my_dbListTables(search_string = "^GR20\\d\\d_PELL_SSL$")
+  tnames <- my_dbListTables(search_string = "^gr20\\d\\d_pell_ssl$")
 
   out <- data.frame()